+++ tests/RollerKostr.Utils.Math.Tests/Implementations/ApproximateGcdCalculatorTests/BinaryApproximateGcdCalculatorTests.cs
@@ -6,7 +6,7 @@ namespace RollerKostr.Utils.Math.Tests.Implementations.ApproximateGcdCalculatorT
 {
     public class BinaryApproximateGcdCalculatorTests
     {
-        private const double FloatingPointTolerance = 0.000001d;
+        private const decimal FloatingPointTolerance = 0.000001m;
         
         [Theory]
         [InlineData( 0,  0, 0)]
@@ -18,10 +18,10 @@ namespace RollerKostr.Utils.Math.Tests.Implementations.ApproximateGcdCalculatorT
         [InlineData( 2,  4, 2)]
         [InlineData( 6, 12, 6)]
         [InlineData(12, 30, 6)]
-        public void Agcd_IfZeroGcdTolerance_ShouldReturnGcd(int a, int b, int expectedAgcd)
+        public void Agcd_IfZeroGcdTolerance_ShouldReturnGcd(decimal a, decimal b, decimal expectedAgcd)
         {
             // Arrange
-            const double gcdTolerancePercent = 0d;
+            const decimal gcdTolerancePercent = 0m;
             
             // Act
             var agcd = ApproximateGcdCalculator.Agcd(gcdTolerancePercent, a, b);
@@ -33,10 +33,10 @@ namespace RollerKostr.Utils.Math.Tests.Implementations.ApproximateGcdCalculatorT
         [Theory]
         [InlineData(1.111, 2.222, 1.111)]
         [InlineData(2.222, 3.333, 1.111)]
-        public void Agcd_ShouldReturnCorrectAgcd(double a, double b, double expectedAgcd)
+        public void Agcd_ShouldReturnCorrectAgcd(decimal a, decimal b, decimal expectedAgcd)
         {
             // Arrange
-            const double gcdTolerancePercent = 0d;
+            const decimal gcdTolerancePercent = 0m;
             
             // Act
             var agcd = ApproximateGcdCalculator.Agcd(gcdTolerancePercent, a, b);
@@ -48,7 +48,8 @@ namespace RollerKostr.Utils.Math.Tests.Implementations.ApproximateGcdCalculatorT
         [Theory]
         [InlineData(1.110, 2.223, 1.110, 1)]
         [InlineData(2.221, 3.334, 1.108, 1)]
-        public void Agcd_ShouldReturnAgcdWithinTolerance(double a, double b, double expectedAgcd, int allowedGcdTolerancePercent)
+        public void Agcd_ShouldReturnAgcdWithinTolerance(decimal a, decimal b, decimal expectedAgcd,
+            decimal allowedGcdTolerancePercent)
         {
             // Arrange
             
@@ -63,7 +64,8 @@ namespace RollerKostr.Utils.Math.Tests.Implementations.ApproximateGcdCalculatorT
         [InlineData(1.110, 2.223, 1)]
         [InlineData(2.221, 3.334, 1)]
         // TODO[mk] generate random data
-        public void Agcd_ShouldReturnValueWhichDivideWithAllowedRemainder(double a, double b, int allowedGcdTolerancePercent)
+        public void Agcd_ShouldReturnValueWhichDivideWithAllowedRemainder(decimal a, decimal b,
+            decimal allowedGcdTolerancePercent)
         {
             // Arrange
             
@@ -71,8 +73,8 @@ namespace RollerKostr.Utils.Math.Tests.Implementations.ApproximateGcdCalculatorT
             var agcd = ApproximateGcdCalculator.Agcd(allowedGcdTolerancePercent, a, b);
 
             // Assert
-            var allowedRemainderA = a * allowedGcdTolerancePercent / 100d;
-            var allowedRemainderB = b * allowedGcdTolerancePercent / 100d;
+            var allowedRemainderA = a * allowedGcdTolerancePercent / 100m;
+            var allowedRemainderB = b * allowedGcdTolerancePercent / 100m;
             var remainderA = a % agcd;
             var remainderB = b % agcd;
             remainderA.Should().BeLessOrEqualTo(allowedRemainderA);
